Clustering QC

Single cells Clustering QC

SCDC_qc(
  sc.eset,
  ct.varname,
  sample,
  scsetname = "Single Cell",
  ct.sub,
  iter.max = 1000,
  nu = 1e-04,
  epsilon = 0.01,
  arow = NULL,
  qcthreshold = 0.7,
  generate.figure = T,
  ct.cell.size = NULL,
  cbPalette = c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
    "#D55E00", "#CC79A7"),
  ...
)

Arguments

sc.eset
ExpressionSet object for single cells

ct.varname
variable name for 'cell type'

sample
variable name for subject/sample

scsetname
the name for the single cell dataset

ct.sub
a subset of cell types that are selected to construct basis matrix

iter.max
the maximum number of iteration in WNNLS

nu
a small constant to facilitate the calculation of variance

epsilon
a small constant number used for convergence criteria

arow
annotation of rows for pheatmap

qcthreshold
the probability threshold used to filter out questionable cells

generate.figure
logical. If generate the heatmap by pheatmap or not. default is TRUE.

ct.cell.size
default is NULL, which means the "library size" is calculated based on the data. Users can specify a vector of cell size factors corresponding to the ct.sub according to prior knowledge. The vector should be named: names(ct.cell.size input) should not be NULL.

Value

a list including: 1) a probability matrix for each single cell input; 2) a clustering QCed ExpressionSet object; 3) a heatmap of QC result.