Clustering QC for single cells from one subject

Clustering QC for single cells from one subject

SCDC_qc_ONE(sc.eset, ct.varname, sample, scsetname = "Single Cell",
  ct.sub, iter.max = 1000, nu = 1e-04, epsilon = 0.01, arow = NULL,
  weight.basis = F, qcthreshold = 0.7, ...)

Arguments

sc.eset
ExpressionSet object for single cells

ct.varname
variable name for 'cell type'

sample
variable name for subject/sample

scsetname
the name for the single cell dataset

ct.sub
a subset of cell types that are selected to construct basis matrix

iter.max
the maximum number of iteration in WNNLS

nu
a small constant to facilitate the calculation of variance

epsilon
a small constant number used for convergence criteria

arow
annotation of rows for pheatmap

qcthreshold
the probability threshold used to filter out questionable cells

Value

a list including: 1) a probability matrix for each single cell input; 2) a clustering QCed ExpressionSet object; 3) a heatmap of QC result.